Fixing issues with pulsed data analysis.
procPlsData had major issues with selecting which groups to include, fixed now.
atplschk now allows time offset to be specified, so that if 'fill' moves the key part of scan around in time, can plot all on top of one another.
Continuing to work on better fitting of echo data.